+++ packages/providers/wagmi/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @betswirl/wagmi-provider
 
+## 0.1.20
+
+### Patch Changes
+
+- Fix leaderboards api version
+
+- Updated dependencies []:
+  - @betswirl/sdk-core@0.1.20
+
+## 0.1.19
+
+### Patch Changes
+
+- Add chain_ids parameter to fetchFreebets function
+
+- Add affiliates & chainIds params to fetchLeaderboards
+
+- Updated dependencies []:
+  - @betswirl/sdk-core@0.1.19
+
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @betswirl/sdk-core@0.1.18
+
 ## 0.1.17
 
 ### Patch Changes