+++ packages/ui-react/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @betswirl/ui-react
 
+## 0.1.16
+
+### Patch Changes
+
+- Fix leaderboards api version
+
+- Updated dependencies []:
+  - @betswirl/wagmi-provider@0.1.20
+  - @betswirl/sdk-core@0.1.20
+
+## 0.1.15
+
+### Patch Changes
+
+- Add chain_ids parameter to fetchFreebets function
+
+- Add affiliates & chainIds params to fetchLeaderboards
+
+- Updated dependencies []:
+  - @betswirl/wagmi-provider@0.1.19
+  - @betswirl/sdk-core@0.1.19
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @betswirl/sdk-core@0.1.18
+  - @betswirl/wagmi-provider@0.1.18
+
 ## 0.1.13
 
 ### Patch Changes